refactor($production): Beans 일급 컬렉션 적용
- bean 들을 생성하고 관리하는 일급 컬렉션 적용
