+++ src/main/java/core/di/factory/BeanFactory.java
@@ -1,23 +1,18 @@
 package core.di.factory;
 
-import static core.di.factory.BeanFactoryUtils.findConcreteClass;
-
 import com.google.common.collect.Maps;
 import core.di.BeanDefinition;
+import core.di.Beans;
 import java.lang.annotation.Annotation;
-import java.lang.reflect.Constructor;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.BeanUtils;
 
 public class BeanFactory {
 
     private static final Logger logger = LoggerFactory.getLogger(BeanFactory.class);
-    private final Map<Class<?>, Object> beans = Maps.newHashMap();
+    private final Beans beans = new Beans();
     private final Map<Class<?>, BeanDefinition> beanDefinitions = Maps.newHashMap();
     private final BeanScanner beanScanner;
 
@@ -27,43 +22,12 @@ public class BeanFactory {
 
     public void initialize(Class<? extends Annotation>... annotations) {
         beanDefinitions.putAll(beanScanner.scanAnnotatedWith(annotations));
-        beanDefinitions.forEach((clazz, beandef) -> beans.put(clazz, instantiate(beandef)));
-    }
-
-    private Object instantiate(BeanDefinition beandef) {
-        if (beans.containsKey(beandef.getBeanClass())) {
-            return beans.get(beandef.getBeanClass());
-        }
-
-        if (beandef.getBeanConstructor() == null) {
-            beans.put(beandef.getBeanClass(), instantiateBean(beandef));
-            logger.debug("Instantiate {} class", beandef.getBeanClass().getName());
-            return beans.get(beandef.getBeanClass());
-        }
-
-        beans.put(beandef.getBeanClass(), instantiateConstructor(beandef.getBeanConstructor()));
-        logger.debug("Instantiate {} class", beandef.getBeanClass().getName());
-        return beans.get(beandef.getBeanClass());
-    }
-
-    private Object instantiateConstructor(Constructor<?> constructor) {
-        Class<?>[] parameterTypes = constructor.getParameterTypes();
-        List<Object> list = new ArrayList<>();
-        for (Class<?> parameterType : parameterTypes) {
-            Class<?> parameterClass = findConcreteClass(parameterType, beanDefinitions.keySet());
-            Object instantiate = instantiate(beanDefinitions.get(parameterClass));
-            list.add(instantiate);
-        }
-        return BeanUtils.instantiateClass(constructor,
-                                          list.toArray());
-    }
-
-    private Object instantiateBean(BeanDefinition beandef) {
-        return BeanUtils.instantiateClass(findConcreteClass(beandef.getBeanClass(), beanDefinitions.keySet()));
+        beans.instantiateBeans(beanDefinitions);
     }
 
     @SuppressWarnings("unchecked")
     public <T> T getBean(Class<T> requiredType) {
+        logger.debug("Get bean {}", requiredType.getName());
         return (T) beans.get(requiredType);
     }
